Move and update KDoc for MARKDOWN_ELEMENT_FILE_NAME
